(#174) Separate console commands description to different file
